# docs/swagger.yaml
...
info:
  contact: {}
  license:
    name: MIT
    url: https://github.com/hello-jiangxiaoyu/QuickAuth/blob/main/LICENSE
  title: Quick Auth 接口文档
  version: "1.0"
paths:
...
  /api/quick/resources/{resourceId}:
    delete:
      parameters:
      - description: resource id
        in: path
        name: resourceId
        required: true
        type: string
      responses:
        "200":
          description: OK
      summary: delete resource
      tags:
      - resource
    get:
      parameters:
      - description: resource id
        in: path
        name: resourceId
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: object
      summary: get resource
      tags:
      - resource
    put:
      parameters:
      - description: resource id
        in: path
        name: resourceId
        required: true
        type: string
      responses:
        "200":
          description: OK
      summary: update resource
      tags:
      - resource
  /api/quick/resources/{resourceId}/json/operations/{operationId}:
    get:
      parameters:
      - description: resource id
        in: path
        name: resourceId
        required: true
        type: string
      - description: json path
        in: query
        name: path
        required: true
        type: string
      - description: operation id
        in: path
        name: operationId
        required: true
        type: string
      responses:
        "200":
          description: OK
      summary: 判断用户当前对JSON资源的操作是否被允许
      tags:
      - auth
  /api/quick/resources/{resourceId}/json/operations/{operationId}/json:
    get:
      parameters:
      - description: resource id
        in: path
        name: resourceId
        required: true
        type: string
      - description: operation id
        in: path
        name: operationId
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: object
      summary: 获取拥有某个操作权限的整个json结构
      tags:
      - auth
  /api/quick/resources/{resourceId}/json/operations/{operationId}/parents/:parentId:
    get:
      parameters:
      - description: resource id
        in: path
        name: resourceId
        required: true
        type: string
      - description: operation id
        in: path
        name: operationId
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: object
      summary: 获取拥有某个操作权限的node列表
      tags:
      - auth
  /api/quick/resources/{resourceId}/json/users/{userId}/roles:
    get:
      parameters:
      - description: resource id
        in: path
        name: resourceId
        required: true
        type: string
      - description: user id
        in: path
        name: userId
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: object
      summary: list resource json user roles
      tags:
      - resource-user
    post:
      parameters:
      - description: resource id
        in: path
        name: resourceId
        required: true
        type: string
      - description: user id
        in: path
        name: userId
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: object
      summary: create resource json user role
      tags:
      - resource-user
  /api/quick/resources/{resourceId}/json/users/{userId}/roles/{roleId}:
    delete:
      parameters:
      - description: resource id
        in: path
        name: resourceId
        required: true
        type: string
      - description: user id
        in: path
        name: userId
        required: true
        type: string
      - description: role id
        in: path
        name: roleId
        required: true
        type: string
      responses:
        "200":
          description: OK
      summary: delete resource json user role
      tags:
      - resource-user
    put:
      parameters:
      - description: resource id
        in: path
        name: resourceId
        required: true
        type: string
      - description: user id
        in: path
        name: userId
        required: true
        type: string
      - description: role id
        in: path
        name: roleId
        required: true
        type: string
      responses:
        "200":
          description: OK
      summary: update resource json user role
      tags:
      - resource-user
  /api/quick/resources/{resourceId}/nodes:
    get:
      parameters:
      - description: resource id
        in: path
        name: resourceId
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: object
      summary: list resource nodes
      tags:
      - resource
    post:
      parameters:
      - description: resource id
        in: path
        name: resourceId
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: object
      summary: create resource node
      tags:
      - resource
  /api/quick/resources/{resourceId}/nodes/{nodeId}:
    delete:
      parameters:
      - description: resource id
        in: path
        name: resourceId
        required: true
        type: string
      - description: node id
        in: path
        name: nodeId
        required: true
        type: string
      responses:
        "200":
          description: OK
      summary: delete resource node
      tags:
      - resource
    get:
      parameters:
      - description: resource id
        in: path
        name: resourceId
        required: true
        type: string
      - description: node id
        in: path
        name: nodeId
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: object
      summary: get resource node
      tags:
      - resource
    put:
      parameters:
      - description: resource id
        in: path
        name: resourceId
        required: true
        type: string
      - description: node id
        in: path
        name: nodeId
        required: true
        type: string
      responses:
        "200":
          description: OK
      summary: update resource node
      tags:
      - resource
  /api/quick/resources/{resourceId}/nodes/{nodeId}/operations/{operationId}:
    get:
      parameters:
      - description: resource id
        in: path
        name: resourceId
        required: true
        type: string
      - description: node id
        in: path
        name: nodeId
        required: true
        type: string
      - description: operation id
        in: path
        name: operationId
        required: true
        type: string
      responses:
        "200":
          description: OK
      summary: 判断用户当前对资源的操作是否被允许
      tags:
      - auth
...
securityDefinitions:
  Login:
    in: header
    name: token
    type: apiKey
swagger: "2.0"